requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.41
psycopg[binary]==3.3.3
itsdangerous==2.2.0
bcrypt>=4.1.0
google-api-python-client>=2.120.0
google-auth>=2.28.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
